refactor(tsconfig-paths): 更新项目配置以支持 tsconfig-paths
- 在 package.json 中添加 tsconfig-paths 依赖
- 更新 start 和 dev 脚本以使用 tsconfig-paths/register
- 移除用户注册相关代码
- 更新评论模型中的文章引用
- 简化用户模型
- 更新路由文件中的导入路径
- 在 tsconfig.json 中添加 baseUrl 配置项
