--- src/models/article.model.ts
@@ -1,10 +1,0 @@
-import { Comment } from './comment.model';
-
-export interface Article {
-  id: number;
-  title: string;
-  slug: string;
-  description: string;
-  comments: Comment[];
-  favorited: boolean;
-}
