+++ src/models/comment.model.ts
@@ -1,9 +1,9 @@
-import { Article } from './article.model';
+import { Post } from './post.model';
 
 export interface Comment {
   id: number;
   createdAt: Date;
   updatedAt: Date;
   body: string;
-  article?: Article;
+  article?: Post;
 }
